configuration for neovim and tweaks to tmux

+++ .config/nvim/init.vim
@@ -8,11 +8,11 @@ set backupdir=~/.vim_temp
 " allow backspacing over everything in insert mode
 set backspace=indent,eol,start
 
-" install vim-plug
-if empty(glob('~/.vim/autoload/plug.vim'))
-  silent !curl -fLo ~/.vim/autoload/plug.vim --create-dirs
+" install vim-plug if not already installed
+if empty(glob('~/.config/nvim/autoload/plug.vim'))
+  silent !curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs
     \ https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-  autocmd VimEnter * PlugInstall | source $MYVIMRC
+  autocmd VimEnter * PlugInstall | source ~/.config/nvim/init.vim
 endif
 
 " ----- Plugins -----
